requirements.txt
# Библиотеки что ставит install_yuko.ps1
# При добавлении нужных просто написать новой строкой

sounddevice
numpy
send2trash
requests
python-dotenv